requirements.txt
asgiref==3.5.2
autopep8==2.0.0
dj-database-url==1.0.0
Django==4.1.3
pycodestyle==2.9.1
python-dotenv==0.21.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
